+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Only write entries that are worth mentioning to users.
 - Lib: Add `Session.dir` property to get the session directory path
 - UI: Improve "Approve for session" experience when there are multiple parallel subagents
 - Wire: Reimplement Wire server mode (which is enabled with `--wire` option)
+- Lib: Rename `ShellApp` to `Shell`, `PrintApp` to `Print`, `ACPServer` to `ACP` and `WireServer` to `WireOverStdio` for better consistency
+- Lib: Rename `KimiCLI.run_shell_mode` to `run_shell`, `run_print_mode` to `run_print`, `run_acp_server` to `run_acp`, and `run_wire_server` to `run_wire_stdio` for better consistency
 
 ## [0.58] - 2025-11-21